init cldf env outside of deployment function, move linking function to existing file
